Simplify shared news parser
