+++ news_common/__init__.py
@@ -4,6 +4,13 @@ import re
 from dataclasses import dataclass, field
 from pathlib import Path
 
+__all__ = [
+    "NewsFragment",
+    "SECTION_ORDER",
+    "is_news_fragment",
+    "parse_sectioned_news",
+]
+
 SECTION_TYPES: dict[str, str] = {
     "Enhancements": "enhancement",
     "Bug fixes": "bugfix",
@@ -15,7 +22,6 @@ SECTION_ORDER: tuple[str, ...] = tuple(SECTION_TYPES)
 
 HEADING_RE = re.compile(r"^(?P<level>#{2,6})\s+(?P<title>.+?)\s*$")
 BULLET_RE = re.compile(r"^\s*[*-]\s+(?P<body>.*\S)\s*$")
-PR_RE = re.compile(r"(?<!\d)#?(?P<number>\d+)(?!\d)")
 
 
 @dataclass(frozen=True)
@@ -31,11 +37,11 @@ class NewsFragment:
         return sum(len(items) for items in self.sections.values())
 
 
-def normalize_heading(value: str) -> str:
+def _normalize_heading(value: str) -> str:
     return " ".join(value.strip().strip("#").strip().casefold().split())
 
 
-SECTION_ALIASES = {normalize_heading(section): section for section in SECTION_ORDER}
+SECTION_ALIASES = {_normalize_heading(section): section for section in SECTION_ORDER}
 
 
 def is_news_fragment(path: str | Path, news_directory: str | Path = "news") -> bool:
@@ -59,17 +65,6 @@ def is_news_fragment(path: str | Path, news_directory: str | Path = "news") -> b
     return relative.suffix in {"", ".md"}
 
 
-def iter_news_fragments(news_directory: str | Path = "news") -> list[Path]:
-    news_directory = Path(news_directory)
-    if not news_directory.is_dir():
-        return []
-    return sorted(
-        path
-        for path in news_directory.iterdir()
-        if path.is_file() and is_news_fragment(path, news_directory)
-    )
-
-
 def parse_sectioned_news(path: str | Path, text: str) -> NewsFragment:
     path = Path(path)
     errors: list[str] = []
@@ -82,7 +77,7 @@ def parse_sectioned_news(path: str | Path, text: str) -> NewsFragment:
     for lineno, line in enumerate(text.splitlines(), start=1):
         if match := HEADING_RE.match(line):
             saw_heading = True
-            title = normalize_heading(match.group("title"))
+            title = _normalize_heading(match.group("title"))
             current_section = SECTION_ALIASES.get(title)
             inside_unknown = current_section is None
 
@@ -170,13 +165,4 @@ def _extract_items(
 
 
 def _is_placeholder(value: str) -> bool:
-    return normalize_heading(value).strip("<>") == "news item"
-
-
-def fragment_mentions_pr(path: str | Path, text: str, pr_number: int | str) -> bool:
-    pr_number = str(pr_number)
-    for value in (Path(path).name, text):
-        for match in PR_RE.finditer(value):
-            if match.group("number") == pr_number:
-                return True
-    return False
+    return _normalize_heading(value).strip("<>") == "news item"
